fixed some endpoints on notification-serv, jwt tokens...

+++ notification-service/src/main/java/com/example/notification_service/Controller/NotificationController.java
@@ -20,34 +20,70 @@ public class NotificationController {
     @Autowired
     private NotificationService notificationService;
 
-    @GetMapping("/user/{userId}")
-    public ResponseEntity<List<Notification>> getUserNotifications(@PathVariable Long userId) {
-        notificationService.getNotificationsByUserId(userId);
-        return ResponseEntity.ok(notificationService.getNotificationsByUserId(userId));
+    @GetMapping("/user")
+    public ResponseEntity<List<Notification>> getUserNotifications(
+            @RequestHeader("Authorization") String token) {
+
+        String username = jwtUtil.extractUsername(token.substring(7));
+
+        return ResponseEntity.ok(notificationService.getNotificationsByUserName(username));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Notification> getNotificationById(@PathVariable Long id) {
+    public ResponseEntity<Notification> getNotificationById(@PathVariable Long id,
+                                                            @RequestHeader("Authorization") String token) {
 
-        return ResponseEntity.ok(notificationService.getNotificationById(id));
+        String username = jwtUtil.extractUsername(token.substring(7));
+
+        Notification notification = notificationService.getNotificationById(id);
+
+        if(!notification.getUsername().equals(username)) {
+            return ResponseEntity.status(403).build();
+
+        }
+
+        return ResponseEntity.ok(notification);
+
     }
 
     @PutMapping("/{id}/mark-as-read")
-    public ResponseEntity<String> markNotificationAsRead(@PathVariable Long id) {
+    public ResponseEntity<String> markNotificationAsRead(@PathVariable Long id,
+                                                         @RequestHeader("Authorization") String token) {
+
+        String username = jwtUtil.extractUsername(token.substring(7));
+        Notification notification = notificationService.getNotificationById(id);
+
+        if(!notification.getUsername().equals(username)) {
+            return ResponseEntity.status(403).body("Unauthorized");
+        }
+
+
         notificationService.markNotificationAsRead(id);
         return ResponseEntity.ok("Notification marked as read!");
 
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteNotification(@PathVariable Long id) {
+    public ResponseEntity<String> deleteNotification(@PathVariable Long id,
+                                                     @RequestHeader("Authorization") String token) {
+
+        String username = jwtUtil.extractUsername(token.substring(7));
+        Notification notification = notificationService.getNotificationById(id);
+
+        if(!notification.getUsername().equals(username)) {
+            return ResponseEntity.status(403).body("Unauthorized");
+        }
+
         notificationService.deleteNotification(id);
         return ResponseEntity.ok("Notification deleted!");
     }
-    @DeleteMapping("/user/{userId}")
-    public ResponseEntity<String> deleteUserNotifications(@PathVariable Long userId) {
-        notificationService.deleteAllNotificationsByUserId(userId);
-        return ResponseEntity.ok("All notifications deleted for user: " + userId);
+    @DeleteMapping("/user")
+    public ResponseEntity<String> deleteUserNotifications(@RequestHeader("Authorization") String token) {
+
+        String username = jwtUtil.extractUsername(token.substring(7));
+
+        notificationService.deleteAllNotificationsByUserName(username);
+        return ResponseEntity.ok("All notifications deleted for user: " + username);
     }
